# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tanim Agricultural Supply Chain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin sees all, buyer sees own, farmer sees their product orders)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/receipt:
    get:
      operationId: getOrderReceipt
      tags: [orders]
      summary: Get PDF receipt URL for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptResponse"
...
components:
  schemas:
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
        subtotal:
          type: number
        unit:
          type: string
      required: [id, product_id, product_name, quantity, unit_price, subtotal, unit]
    Order:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        buyer_name:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        payment_method:
          type: string
          enum: [cod, online]
        payment_status:
          type: string
          enum: [pending, paid, failed, refunded]
        total_amount:
          type: number
        shipping_address:
          type: string
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, user_id, buyer_name, status, payment_method, payment_status, total_amount, shipping_address, created_at, updated_at]
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [orders, total, page, limit]
    CreateOrderRequest:
      type: object
      properties:
        shipping_address:
          type: string
        payment_method:
          type: string
          enum: [cod, online]
        notes:
          type: string
      required: [shipping_address, payment_method]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        notes:
          type: string
      required: [status]
    ReceiptResponse:
      type: object
      properties:
        receipt_url:
          type: string
        order_id:
          type: integer
        message:
          type: string
      required: [receipt_url, order_id, message]
...