# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tanim Agricultural Supply Chain API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderPaymentStatus } from "./orderPaymentStatus";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  user_id: number;
  buyer_name: string;
  status: OrderStatus;
  payment_method: OrderPaymentMethod;
  payment_status: OrderPaymentStatus;
  total_amount: number;
  shipping_address: string;
  notes?: string | null;
  created_at: Date;
  updated_at: Date;
}